Abstracted out the nerve builder to a separate class
The previous method `_create_links` is now implemented by `.nerve.GraphNerve`, and will return a list of simplices as well as the neighbor graph needed for `links`.
By returning a full simplex list, alternative nerve methods that create the entire Cech complex of the refined pulled back cover will be easier to implement and integrate, and with them integration of homology (persistent or otherwise) will be easier.