Enable precomputed similarity matrices
Clustering functions such as `DBSCAN` accept `metric='precomputed'`, which requires a square distance (similarity) matrix to be passed as, in this case, `inverse_X`. However, the mapper isn't aware when
precomputed matrices are being passed in, so when it slices according to the filter function values and hypercubes, the square-ness is un-squared.

This PR adds a kludgey way to tell the mapper that a precomputed matrix
is being passed in. When set to `True`, the slice will feed a square
matrix to `clusterer.fit()`